separate usage of words Laser and Profile in class names
