+++ lasy/profiles/profile.py
@@ -1,7 +1,7 @@
 import numpy as np
 import scipy.constants as scc
 
-class LaserProfile(object):
+class Profile(object):
     """
     Base class for all laser profiles.
